clean up test folder
