+++ source/geh_calculated_measurements/tests/electrical_heating/conftest.py
@@ -3,14 +3,8 @@ from pathlib import Path
 from typing import Generator
 
 import pytest
-from geh_common.pyspark.read_csv import read_csv_path
-from geh_common.testing.delta_lake.delta_lake_operations import create_database, create_table
 from pyspark.sql import SparkSession
 
-from geh_calculated_measurements.electrical_heating.domain import time_series_points_v1
-from geh_calculated_measurements.electrical_heating.infrastructure import (
-    MeasurementsGoldDatabaseDefinition,
-)
 from tests import PROJECT_ROOT
 from tests.electrical_heating.testsession_configuration import TestSessionConfiguration
 
@@ -40,29 +34,3 @@ def contracts_path() -> str:
 def test_session_configuration(tests_path: str) -> TestSessionConfiguration:
     settings_file_path = Path(tests_path) / "testsession.local.settings.yml"
     return TestSessionConfiguration.load(settings_file_path)
-
-
-@pytest.fixture(scope="session")
-def test_files_folder_path(tests_path: str) -> str:
-    return f"{tests_path}/utils/test_files"
-
-
-@pytest.fixture(scope="session")
-def seed_gold_table(spark: SparkSession, test_files_folder_path: str) -> None:
-    create_database(spark, MeasurementsGoldDatabaseDefinition.DATABASE_NAME)
-
-    create_table(
-        spark,
-        database_name=MeasurementsGoldDatabaseDefinition.DATABASE_NAME,
-        table_name=MeasurementsGoldDatabaseDefinition.TIME_SERIES_POINTS_NAME,
-        schema=time_series_points_v1,
-        table_location=f"{MeasurementsGoldDatabaseDefinition.DATABASE_NAME}/{MeasurementsGoldDatabaseDefinition.TIME_SERIES_POINTS_NAME}",
-    )
-
-    file_name = f"{test_files_folder_path}/{MeasurementsGoldDatabaseDefinition.DATABASE_NAME}-{MeasurementsGoldDatabaseDefinition.TIME_SERIES_POINTS_NAME}.csv"
-    time_series_points = read_csv_path(spark, file_name, time_series_points_v1)
-    time_series_points.write.saveAsTable(
-        f"{MeasurementsGoldDatabaseDefinition.DATABASE_NAME}.{MeasurementsGoldDatabaseDefinition.TIME_SERIES_POINTS_NAME}",
-        format="delta",
-        mode="overwrite",
-    )
