+++ source/geh_calculated_measurements/tests/electrical_heating/job_tests/conftest.py
@@ -1,0 +1,43 @@
+import pytest
+from geh_common.pyspark.read_csv import read_csv_path
+from geh_common.testing.delta_lake.delta_lake_operations import create_database, create_table
+from pyspark.sql import SparkSession
+
+from geh_calculated_measurements.capacity_settlement.contracts.measurements_gold.time_series_points_v1 import (
+    time_series_points_v1,
+)
+
+
+@pytest.fixture(scope="session")
+def test_files_folder_path(tests_path: str) -> str:
+    return f"{tests_path}/job_tests/test_files"
+
+
+@pytest.fixture(scope="session")
+def seed_gold_table(spark: SparkSession, test_files_folder_path: str) -> None:
+    create_database(spark, MeasurementsGoldDatabaseDefinition.DATABASE_NAME)
+
+    create_table(
+        spark,
+        database_name=MeasurementsGoldDatabaseDefinition.DATABASE_NAME,
+        table_name=MeasurementsGoldDatabaseDefinition.TIME_SERIES_POINTS_NAME,
+        schema=time_series_points_v1,
+        table_location=f"{MeasurementsGoldDatabaseDefinition.DATABASE_NAME}/{MeasurementsGoldDatabaseDefinition.TIME_SERIES_POINTS_NAME}",
+    )
+
+    file_name = f"{test_files_folder_path}/{MeasurementsGoldDatabaseDefinition.DATABASE_NAME}-{MeasurementsGoldDatabaseDefinition.TIME_SERIES_POINTS_NAME}.csv"
+    time_series_points = read_csv_path(spark, file_name, time_series_points_v1)
+    time_series_points.write.saveAsTable(
+        f"{MeasurementsGoldDatabaseDefinition.DATABASE_NAME}.{MeasurementsGoldDatabaseDefinition.TIME_SERIES_POINTS_NAME}",
+        format="delta",
+        mode="overwrite",
+    )
+
+
+@pytest.fixture(scope="session")
+def job_environment_variables(test_files_folder_path) -> dict:
+    return {
+        "CATALOG_NAME.name": "spark_catalog",
+        "TIME_ZONE": "Europe/Copenhagen",
+        "ELECTRICITY_MARKET_DATA_PATH": test_files_folder_path,
+    }
