+++ source/geh_calculated_measurements/tests/electrical_heating/job_tests/test_files/consumption_metering_point_periods_v1.csv
